Create SubscriptionUtils
