+++ app/src/main/java/com/tehmou/rxbookapp/viewmodels/BookViewModel.java
@@ -83,5 +83,4 @@ public class BookViewModel {
                     }
                 });
     }
-
 }
